one-off scripts that perform migration on a temp ec2 instance
